+++ bindings/mnt/pyfiction/src/pyfiction/algorithms/physical_design/apply_gate_library.cpp
@@ -3,9 +3,9 @@
 
 #include <fiction/algorithms/physical_design/apply_gate_library.hpp>
 #include <fiction/technology/inml_topolinano_library.hpp>
-#include <fiction/technology/molecular_qca_library.hpp>
 #include <fiction/technology/qca_one_library.hpp>
 #include <fiction/technology/sidb_bestagon_library.hpp>
+#include <fiction/technology/sim7_mol_library.hpp>
 #include <fiction/traits.hpp>
 
 #include <fmt/format.h>
@@ -46,7 +46,7 @@ void apply_fcn_gate_library(nanobind::module_& m, const std::string& lib_name)
 void apply_gate_library(nanobind::module_& m)
 {
     detail::apply_fcn_gate_library<fiction::qca_one_library, py_cartesian_gate_layout>(m, "qca_one");
-    detail::apply_fcn_gate_library<fiction::molecular_qca_library, py_cartesian_gate_layout>(m, "mol_qca");
+    detail::apply_fcn_gate_library<fiction::sim7_mol_library, py_cartesian_gate_layout>(m, "sim7_mol");
     detail::apply_fcn_gate_library<fiction::inml_topolinano_library, py_shifted_cartesian_gate_layout>(m, "topolinano");
     detail::apply_fcn_gate_library<fiction::sidb_bestagon_library, py_hexagonal_gate_layout>(m, "bestagon");
 }