+++ cli/cmd/technology/src/cell.cpp
@@ -4,7 +4,7 @@
 
 #include "cmd/technology/include/cell.hpp"
 
-#include "fiction/technology/molecular_qca_library.hpp"
+#include "fiction/technology/sim7_mol_library.hpp"
 #include "stores.hpp"  // NOLINT(misc-include-cleaner)
 
 #include <fiction/algorithms/physical_design/apply_gate_library.hpp>
@@ -31,7 +31,7 @@ cell_command::cell_command(const environment::ptr& e) :
         command(e, "Compiles the current gate layout in store down to a cell-level layout. A gate library must be "
                    "specified in order to instruct the algorithm how to map gate tiles to cell blocks.")
 {
-    add_option("--library,-l", library, "Gate library to use for mapping {QCA-ONE, molQCA, ToPoliNano, Bestagon}",
+    add_option("--library,-l", library, "Gate library to use for mapping {QCA-ONE, SIM7-MOL, ToPoliNano, Bestagon}",
                true);
 }
 
@@ -66,15 +66,15 @@ void cell_command::execute()
             store<fiction::cell_layout_t>().extend() = *result;
         }
     }
-    else if (library == "MOLQCA" || library == "MOL-QCA" || library == "MOL QCA")
+    else if (library == "SIM7" || library == "SIM7MOL" || library == "SIM7-MOL" || library == "SIM7_MOL")
     {
-        const auto apply_mol_qca = [](auto&& lyt_ptr)
+        const auto apply_sim7_mol = [](auto&& lyt_ptr)
         {
             return std::make_shared<fiction::mol_qca_cell_clk_lyt>(
-                fiction::apply_gate_library<fiction::mol_qca_cell_clk_lyt, fiction::molecular_qca_library>(*lyt_ptr));
+                fiction::apply_gate_library<fiction::mol_qca_cell_clk_lyt, fiction::sim7_mol_library>(*lyt_ptr));
         };
 
-        const auto visitor = [&apply_mol_qca](auto&& source) { return std::visit(apply_mol_qca, source); };
+        const auto visitor = [&apply_sim7_mol](auto&& source) { return std::visit(apply_sim7_mol, source); };
 
         if (const auto result = apply_with_error_handling(visitor, s.current()); result.has_value())
         {