+++ test/algorithms/physical_design/apply_gate_library.cpp
@@ -16,12 +16,12 @@
 #include <fiction/layouts/coordinates.hpp>
 #include <fiction/layouts/gate_level_layout.hpp>
 #include <fiction/layouts/tile_based_layout.hpp>
-#include <fiction/technology/molecular_qca_library.hpp>
 #include <fiction/technology/qca_one_library.hpp>
 #include <fiction/technology/sidb_bestagon_library.hpp>
 #include <fiction/technology/sidb_defect_surface.hpp>
 #include <fiction/technology/sidb_defects.hpp>
 #include <fiction/technology/sidb_on_the_fly_gate_library.hpp>
+#include <fiction/technology/sim7_mol_library.hpp>
 #include <fiction/traits.hpp>
 #include <fiction/types.hpp>
 #include <fiction/utils/truth_table_utils.hpp>
@@ -606,7 +606,7 @@ TEST_CASE("Apply molecular QCA gate library end-to-end", "[apply-gate-library]")
 
     const auto layout = blueprints::and_or_inv_gate_layout<gate_layout>();
 
-    const auto cell_layout = apply_gate_library<mol_qca_cell_clk_lyt, molecular_qca_library>(layout);
+    const auto cell_layout = apply_gate_library<mol_qca_cell_clk_lyt, sim7_mol_library>(layout);
 
     CHECK(cell_layout.num_cells() > 0u);
 }